requirements.txt
pip install django
djangorestframework
psycopg2
cryptography
djangorestframework-jwt
pyotp
backports.pbkdf2
binascii
pycryptodome
twillo
pycrypto
django-cors-headers